move query param check to middleware and config
